requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
polars>=0.20.0
pyarrow>=14.0.1
fastparquet>=2023.10.1
dask>=2024.1.0        # For parallel computing
distributed>=2024.1.0  # Dask's distributed scheduler
dask[complete]>=2024.1.0  # All Dask dependencies including diagnostics

# Data processing and validation
pvlib>=0.10.0  # For solar position calculations
geopy>=2.4.0   # For geocoding
python-dateutil>=2.8.2

# Database and storage
sqlalchemy>=2.0.0
pyodbc>=5.0.1
duckdb>=0.9.2

# API and web
requests>=2.31.0
httpx>=0.25.0
aiohttp>=3.9.0
openmeteo-requests>=1.5.0  # For weather data API
requests-cache>=1.2.1      # For caching API requests
retry-requests>=2.0.0      # For handling API retries

# Machine learning
scikit-learn>=1.3.0
prophet>=1.1.4
xgboost>=2.0.0
mlflow>=2.8.0

# Visualization
streamlit>=1.29.0
plotly>=5.18.0
matplotlib>=3.8.0

# LLM and AI
langchain>=0.1.0
openai>=1.3.0
faiss-cpu>=1.7.4
chromadb>=0.4.18
langgraph>=0.0.15  # For agent orchestration

# Testing and development
pytest>=7.4.0
black>=23.9.0
isort>=5.12.0
flake8>=6.1.0

# Logging and monitoring
python-json-logger>=2.0.7
prometheus-client>=0.17.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
pyyaml>=6.0.1 